+++ cli/cli.hs
@@ -41,7 +41,7 @@ import JSONL (jsonlBuilder, jsonLine)
 
 import EVM (initialContract, abstractContract, makeVm, defaultVMOpts)
 import EVM.ABI (Sig(..))
-import EVM.Dapp (dappInfo, DappInfo, emptyDapp)
+import EVM.Dapp (dappInfo, DappInfo, emptyDapp, TestMethodInfo(..))
 import EVM.Expr qualified as Expr
 import EVM.Concrete qualified as Concrete
 import EVM.Fetch qualified as Fetch
@@ -193,8 +193,8 @@ data TestOptions = TestOptions
   , rpc           ::Maybe URL
   , number        ::Maybe W256
   , coverage      ::Bool
-  , match         ::Maybe String
-  , prefix        ::String
+  , match         ::Maybe Text
+  , prefix        ::Text
   , ffi           ::Bool
   }
 
@@ -431,7 +431,7 @@ equivalence eqOpts cOpts = do
       (False, False) -> putStrLn "   \x1b[32m[PASS]\x1b[0m Contracts behave equivalently"
       (True, _)      -> putStrLn "   \x1b[31m[FAIL]\x1b[0m Contracts do not behave equivalently"
       (_, True)      -> putStrLn "   \x1b[31m[FAIL]\x1b[0m Contracts may not behave equivalently"
-    liftIO $ printWarnings Nothing eq.partials (map fst eq.res) "the contracts under test"
+    liftIO $ printWarnings Nothing mempty eq.partials (map fst eq.res) "the contracts under test"
     case any (isCex . fst) eq.res of
       False -> liftIO $ do
         when anyIssues exitFailure
@@ -548,7 +548,7 @@ symbCheck cFileOpts sOpts cExecOpts cOpts = do
                  , ""
                  ] <> fmap (formatCex (fst calldata) Nothing) cexs
         liftIO $ T.putStrLn $ T.unlines counterexamples
-        liftIO $ printWarnings Nothing expr res "symbolically"
+        liftIO $ printWarnings Nothing mempty expr res "symbolically"
         showExtras solvers sOpts calldata expr
         liftIO exitFailure
 
@@ -864,8 +864,7 @@ unitTestOptions testOpts cOpts solvers buildOutput = do
     , maxIter = parseMaxIters cOpts.maxIterations
     , askSmtIters = cOpts.askSmtIterations
     , smtTimeout = Just cOpts.smtTimeout
-    , match = T.pack $ fromMaybe ".*" testOpts.match
-    , prefix = T.pack testOpts.prefix
+    , methodFilter = \(TestMethodInfo _ (Sig name _)) -> testOpts.prefix `T.isPrefixOf` name && maybe True (\match -> regexMatches match name) testOpts.match
     , testParams = params
     , dapp = srcInfo
     , ffiAllowed = testOpts.ffi